Introduce OsvProcessBuilder.
OsvProcessBuilder is a replacement for the standard java ProcessBuilder.
Because we can not spawn new processes in OSv we could not start new JVMs.
OsvProcessBuilder starts new JVMs as threads instead of as processes. This
allows us to start new JVMs inside OSv and greatly improves the versatility
of OSv usecases.